job:  #3  output routine mixin
This commit factors out the JSON production and pretty print
routines into separate files using a "mixin" pattern of inheritance.
The intent is to keep the guts (data model) of the job, sequence,
event definitions and controls contained.  And allow for purpose-
specific output routines to live in single files.