job:  #3  deal with merge and reorganize
This commit plays a merge point only once allowing the branches
to complete before proceeding from the merge point.  This is done
by comparing the defined previous events (expected) with those
actually seen so far.
The test input files are synchronized with those in the Protocol Verifier.
The regression scripts are moved into the tests folder.
A class diagram is shown in the README.