job:  #MUN2-68  better play
This commit plays loops with previousEventIds and forks better with
previousEventIds (instances).
An AuditEvent now carries a scope.  So it knows how deeply it is nested.
This is used in the search backwards for an AND constraint at the merge
points.  It causes the search to avoid more deeply nested constraints
behind it.
Loops can now loop around a single event.