job:  #MUN2-94  support merge count play
This is handled similarly to an AND fork.  When the user of a merge
count is detected, the next event is not played until the visit count
equals the merge count.  This allows the other instance forks to play
first.  Then, the merge point gets all of the previous event Ids
appended into the merge user event.