job:  #MUN2-117  proposed simple unhappy play
This is a first pass and simple mechanism for playing with play in the
context of critical and unhappy events.  It is likely to need to be
changed, but I will commit it as potentially useful in the short term.
When an audit event definition is encountered with IsCritical, a coin
is tossed.  If heads, play the critical audit event like normal.
If tails, play an unhappy event in place of the critical event and
continue.  Continuing will play other audit events that follow the
critical event.  This may be dubious, but it is what it is for now.