job:  #MUN2-135  problem determination
There is a flaw in the Pathway assembly centered on linking alternatives
in parallel Tines of an AND Fork.  The present method of determining an
"upstream" alternative scopes upwards only and not to sibling scopes in
the multiple Tines of an AND Fork.

This is not a critical issue.  There is only the possibly of less than 100%
coverage of job definitions containing XOR Forks nested inside sibling Tines
of AND Forks.