job:  #63  support if without else
This commit makes a change to allow an if statement with only the then
side and no else side.  This amounts to a fork with an empty tine.
The empty tine serves as a short-cut from the top of the fork to the
bottom.  It implies that an event after the fork can have the event
at the start of the fork as a previous event.