job:  #71  more robust start detection
With nested features (Fork), the current method of detecting a
start event was lacking.  This commit switches to detecting and
linking start events after the sequence is completely parsed.
This detects start events that may be further along in the list
of PLUS/PUML statements.
The deeply_nested_XOR.puml is added to the munin regression
test cases.